# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Red Light Green Light Platform API",
    "version": "1.0.0",
    "description": "Unified public API served through the gateway. Generated from per-service controller annotations — do not edit by hand (run `npm run generate:openapi`).",
    "contact": {
      "name": "Red Light Green Light Support"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "API Gateway"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TestCaseInput": {
        "properties": {
          "customFieldValues": {
            "additionalProperties": true,
            "type": "object"
          },
          "description": {
            "maxLength": 2000,
            "type": "string"
          },
          "expectedResult": {
            "maxLength": 1000,
            "type": "string"
          },
          "preconditions": {
            "maxLength": 1000,
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "projectId": {
            "type": "integer"
          },
          "sectionId": {
            "type": "integer"
          },
          "steps": {
            "description": "Ordered list of step objects (at least one required)",
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "suiteId": {
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": ["title", "steps", "sectionId", "suiteId", "projectId"],
        "type": "object"
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/testcases/cases": {
      "get": {
        "parameters": [
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "suiteId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sectionId",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of test cases"
          }
        },
        "summary": "List test cases in a project (optionally filtered by suite or section)",
        "tags": ["TestCases"]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestCaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created test case"
          }
        },
        "summary": "Create a new test case",
        "tags": ["TestCases"]
      }
    },
    "/api/v1/testcases/cases/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Soft-delete a test case",
        "tags": ["TestCases"]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test case"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Get a single test case by ID",
        "tags": ["TestCases"]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-project-id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestCaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated test case"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Version conflict (optimistic concurrency)"
          }
        },
        "summary": "Update a test case",
        "tags": ["TestCases"]
      }
    },
...
  }
}